Changed order of rules. Moved "Set Release Tier" to a higher priority than "Preliminary Set Out Year Release Volume to 8.23MAF". This doesn't cause problems in the normal Powell policy group since the rule "Set Release Tier" doesn't have the EOCY Powell Pool Elevation. Therefore, the  rule "Set Release Tier" is skipped until the rule "Preliminary Set Out Year Release Volume to 8.23MAF" has executed successfully. The rule priorities were changed to make the ruleset more efficiently.
The priority issue was causing incorrectly calculated Release Tiers in the post-LBDV rules. When the ruleset got to the rule "Set Release Tier post-LBDV" the release for the water year was still set to a higher TARV (9.0MAF). This caused the rules to calculate the wrong release. MTOM then reset the TARV to 8.23MAF in "Preliminary Set Out Year Release Volume to 8.23MAF post-LBDV". Switching the order of the rules allows for the TARV to be reset to 8.23MAF before calculating the release tier.