Edited function: Volume DeviationFromEOWYTarget
This function calculates a VolumeDeviation in the rule “Refine Equalization Release Volume”.  This function needs to account for carryover volume that is created for the current year; without accounting for carryover, MTOM continuously increase the TARV and eventually aborts. Added two 'WITH' statements to use for current and previous year's carryover volume.