Function: GetPowellMonthlyReleases
S.Baker, 2018/10/17:
(1) 'AnnualCarryoverVolume' - Function needs to know the carryover volume no matter the equalization tier.
(2) 'RemainingMonthlyReleases' - The ‘CalculatePowellUEBToEqualizePowellMonRel’ function uses ‘Powell.Peak Flow’ which isn’t solved for until Powell has successfully dispatched. Therefore, the statement setting the ‘RemainingMonthlyRelease’ checks to see of Powell.Outflow is NaN. This normally occurs when the TARV is initially set to 8.23 MAF. Once Powell resets the TARV, the function will re-execute and will enter the function ‘CalculatePowellUEBToEqualizePowellMonRel’ if Powell is equalizing with an April adjustment.
(3) IF statement for ‘RemainingMonthlyReleases’ - A new conditions was added to ensure the model didn’t enter this part of the function unless it was not an April Adjustment. Originally, the model would solve for carryover, but would spread the carryover out by spilling water out of Powell in this part of the function.