Rules: Preliminary Set WY 1 Release Volume To 8.23MAF (and post-LBDV rule)
Changed column number in releaseTable[AnnualVolume, Ncol - 3]
MTOM was failing in testbed runs in this rule in high inflow years when Powell was trying to release a large volume. In the historical run of 1983-06, the model aborted since it sets the initial TARV to 30 MAF based on the condition releaseTable[AnnualVolume, Ncol – 2].  This caused the model to look for too high of upper bound in the function ‘CalculatePowellMonthlyRelease’.  The upper bound was being reset in function 'FindRemainingReleaseUpperBoundColumn' by adding + 1 index which cause model to look for table index that didn’t exist. When we use Ncol – 3, the TARV is initially set to 20 MAF and MTOM runs successfully finding a realistic TARV (18 MAF).