Rule edit: Mead Flood Control Release
Add constraint within IF statement to check if Mead.Outflow is greater than 0  (non-negative) this rule was setting Mead outflow during very low scenarios in which flood control rules should not be executing.

Excel IC file edits - added TARV
The IC files were updated to include a TARV for 2020 which was needed for simulations starting in August and September.