Powell.DataPowellMonthlyReleaseTable - add column
Added one higher column for the rules to iterate off of. An error was occurring in the rules that caused MTOM to abort when Powell was in UEBT. The rules tried to release 9.0MAF from Oct-Mar, and then the remaining volume (18MAF) through the rest of the WY. This was causing the function to look for a column that didnt exist in the table. I added on column following the same release pattern for 75MF. The rules dont set final releases based on these columns because of constraints at Powell, but it allows the rules to solve and move on to adjust the release to a realistic value.