Function: GetPowellReleasesCorrectedForBypass
Had previously edited the assignment of addCarryoverIfAny in the statement Min(PreviousCarryover, index(2)). The index(2) part was necessary in case it was removed in the previous timestep by the GetPowellMonthlyReleases function. Added it back in and the 2019-12-avg-Trace6 scenario ran fine (it was originally calculating carryover when there was none).