Edit Powell rules for "Set Upper Balancing Release Volume [post-LBDV]".
I found an error in Lake Powell rules for 'Set Upper Balancing Release Volume' (and post-LBDV). When the Upper Elevation Balancing tier projects the EOWY Powell PE above the Equalization line, the model calculated Powell outflows less than 7 maf. Based on the interim guidelines for UEB (6.B.3), equalization is supposed to govern this situation, but with a minimum release of 8.23 maf. I had to make a minor change to the two rules affected to fix this issue. This including adding a Max() constraint with 8.23 maf.