mdl -  Change Max Vol to dummy value for extreme case when the reservoir is really high and the flow to get it back below the top of the PE-Vol curve max value exceeds the safe channel capacity. In this extreme case calculating a safe channel capacity release using the dummy PE is needed
rls -  Added NOT HasRuleFiredSuccessfully() on 'Adjust Daily for August Base Flow' rule. What happens is the Rule 95 (August Daily) is dependent on FlamingGorge.Outflow which has been solved previously by Rule 99 (Baseflow).  Rule 95 creates a daily flow list which is summed up by Rule 3 (Sum Daily) which sets FlamingGorge.Outflow which causes 95, which depends on this, to go back on the que. However this time 95 runs the Outflow is different enough that it solves a different way (first time requires August down ramping, second time does not) then we go back to rule 93 which sums something new and ad nauseam